Add new modules in modules directory
