+++ modules/new-book.js
@@ -1,0 +1,11 @@
+const booksSection = document.querySelector('.display-books');
+const formSection = document.querySelector('.form-section');
+const contactSection = document.querySelector('.contact-section');
+
+function showForm() {
+  booksSection.classList.add('hide');
+  formSection.classList.remove('hide');
+  contactSection.classList.add('hide');
+}
+
+export default showForm;
